+++ NEWS.md
@@ -1,5 +1,7 @@
 # RPresto 1.3.2.9000
 
+- Send headers with all http requests (#103).
+
 # RPresto 1.3.2
 
 - Use the new delayed S3 registration mechanism in 3.6.0 for dplyr method implementations.